backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# FHIR Client
fhirclient==4.1.0
fhirpy==2.0.5

# Azure SDK
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
azure-keyvault-secrets==4.7.0

# OpenAI (Azure OpenAI compatible)
openai==1.7.2

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-json-logger==2.0.7

# Date/Time
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
respx==0.21.1
faker==21.0.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Monitoring
prometheus-client==0.19.0
